requirements.txt
#general module
django >4.2 ,<4.3
python-decouple
pillow
djangorestframework

#third party module
drf-yasg[validation]
markdown
django-filter


#refactoring and tests
black